# lib/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Bizny API - Productivity and venture coordination platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistration"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
components:
  schemas:
...
    UserRegistration:
      type: object
      required: [name, email, whatsapp, country, industry, role]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        whatsapp:
          type: string
        country:
          type: string
        industry:
          type: string
        role:
          type: string
        skills:
          type: array
          items:
            type: string
        interests:
          type: array
          items:
            type: string
    LoginInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, email, country, industry, role, verificationStatus, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        whatsapp:
          type: ["string", "null"]
        country:
          type: string
        industry:
          type: string
        role:
          type: string
        bio:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        interests:
          type: array
          items:
            type: string
        verificationStatus:
          type: string
          enum: [unverified, pending, verified]
        avatarUrl:
          type: ["string", "null"]
        isBusiness:
          type: boolean
        businessName:
          type: ["string", "null"]
        businessRegistrationNumber:
          type: ["string", "null"]
        stateCity:
          type: ["string", "null"]
        subIndustries:
          type: array
          items:
            type: string
        primaryProducts:
          type: array
          items:
            type: string
        services:
          type: array
          items:
            type: string
        phone:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        publicSlug:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...